tasks: pkg_build_target: pass source package name to dck-buildpackage
dck-buildpackage tries to guess the source package name - the name
is used as subdirectory within the target apt repository. if there's
no debian/control file, it tries to create it by calling debian/rules.

this is the unpleasant side effect that debian/rules is currently done
on the host, instead of within the build container (could be considered
a bug in dck-buildpackage). for some packages this causes trouble:
eg. the in the linux kernel, the .config file is rewritten.

the easiest way for now is just passing the source package, so
dck-buildpackage doesn't need to guess, therefore doesn't need to
call debian/rules on the host.