metux: lambdadict: add defaults support
Adding an internal dict for defaults, where we can fallback to if a
requested key is missing. The idea is strictly separating defaults
from actually defined values, so we don't need to care about load
ordering, can change defaults anytime or remove keys and so let the
defaults take effect again.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>