metux: specobject: helper class for mapping sub-dicts into own namespace
Introducing an helper SpecObject which accesses all elements within
a sub-dict / namespace of another SpecObject. This can be used as an
intrinsic by other SpecObject, eg. if different sub-dicts (or many
of their fields) from somewhere else shall be visible under a canonical
path, based on certain circumstances. Somewhat in the spirit of bind-
mounts in filesystem.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>