deb_autopkg: cmd: return 0 on success
The cmd functions are supposed to return an process exit code, so the
calling command script can just do sys.exit(func()). But on success
(no exception) they currently just return the Builder method's retval,
which is true or false depending on whether one of the task had actually
run (not skipped).

Obviously that's wrong: the command scripts will return nonzero exit code,
indicating an error, even if it was sucess.

Fixing this by returning 0 instead of Builder's retval in cmd wrapper, when
no exception happened. (on exception, we already correctly return nonzero)

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>